refactor chagpt module to openai module
